+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### NEXT
 
-- Node: Make `RtpCodecCapability.preferredPayloadType` mandatory and add `RouterRtpCodecCapability` in which `preferredPayloadType` is optional ([PR #15XX](https://github.com/versatica/mediasoup/pull/15XX)).
+- Node: Make `RtpCodecCapability.preferredPayloadType` mandatory and add `RouterRtpCodecCapability` type (in which `preferredPayloadType` is optional) and `RouterRtpCapabilities` type ([PR #15XX](https://github.com/versatica/mediasoup/pull/15XX)).
 
 ### 3.17.1